# agent/common/openapi_spec.yaml
openapi: 3.0.3
info:
  title: Recipe Service API
  description: |
    레시피 추천 및 재료 관리 서비스 API
    
    ## 인증 방식
    - Cookie 기반 JWT 인증 (HttpOnly, Secure, SameSite=Strict)
    - 소셜 로그인: 백엔드 주도 OAuth. 클라이언트는 GET /api/v1/auth/{provider} 진입만 하며, 이후 리다이렉트·토큰 교환·JWT 발급은 백엔드에서 처리
    
    ## 주요 기능
    - 소셜 로그인 (카카오, 구글, 네이버)
    - 레시피 조회 및 검색
    - 사용자 재료 관리
    - AI 챗봇 레시피 추천
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.example.com
    description: Production server
  - url: https://staging-api.example.com
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/users/me/inventory/ingredients:
    get:
      tags:
        - Inventory
      summary: 내 보관함 조회
      description: 사용자의 보유/관심 재료 목록을 함께 조회
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
      responses:
        '200':
          description: 보관함 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/inventory/ingredients/owned:
    put:
      tags:
        - Inventory
      summary: 내 보유 재료 업데이트 (전체 교체)
      description: 보유 재료 목록 전체 교체
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ownedIngredientIds
              properties:
                ownedIngredientIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '200':
          description: 업데이트 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Inventory
      summary: 내 보유 재료 추가
      description: 보유 재료 목록에 재료 ID 추가
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ownedIngredientIds
              properties:
                ownedIngredientIds:
                  type: array
                  items:
                    type: integer
                    format: int64
      responses:
        '201':
          description: 추가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/inventory/ingredients/owned/{ingredientId}:
    delete:
      tags:
        - Inventory
      summary: 내 보유 재료 삭제
      description: 보유 재료 목록에서 특정 재료 삭제
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
        - name: ingredientId
          in: path
          required: true
          description: 재료 ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/inventory/ingredients/favorites:
    put:
      tags:
        - Inventory
      summary: 내 관심 재료 업데이트 (전체 교체)
      description: 관심 재료 목록 전체 교체
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - favoriteIngredientIds
              properties:
                favoriteIngredientIds:
                  type: array
                  items:
                    type: integer
                    format: int64
            example:
              favoriteIngredientIds: [1, 5, 12, 23]
      responses:
        '200':
          description: 업데이트 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Inventory
      summary: 내 관심 재료 추가
      description: 관심 재료 목록에 재료 ID 추가 (기존 목록에 병합)
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - favoriteIngredientIds
              properties:
                favoriteIngredientIds:
                  type: array
                  items:
                    type: integer
                    format: int64
            example:
              favoriteIngredientIds: [1, 5]
      responses:
        '201':
          description: 추가 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/users/me/inventory/ingredients/favorites/{ingredientId}:
    delete:
      tags:
        - Inventory
      summary: 내 관심 재료 삭제
      description: 관심 재료 목록에서 특정 재료 제거
      parameters:
        - name: accessToken
          in: cookie
          required: true
          description: JWT 액세스 토큰
          schema:
            type: string
          example: "{{apiKey}}"
        - name: ingredientId
          in: path
          required: true
          description: 제거할 재료 ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: 삭제 성공
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: JWT 토큰이 포함된 HttpOnly 쿠키
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: 내부 API 인증 키
  schemas:
...
    InventoryEntry:
      type: object
      required:
        - id
        - name
        - categoryId
      properties:
        id:
          type: integer
          format: int64
          description: 재료 ID
        name:
          type: string
          description: 재료명
        categoryId:
          type: integer
          format: int64
          nullable: true
          description: 재료 카테고리 ID (마스터에 없는 ID는 null)
    InventoryResponse:
      type: object
      properties:
        ownedIngredients:
          type: array
          description: 보유 재료 (id, name, categoryId)
          items:
            $ref: '#/components/schemas/InventoryEntry'
        favoriteIngredients:
          type: array
          description: 관심 재료 (id, name, categoryId)
          items:
            $ref: '#/components/schemas/InventoryEntry'
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              nullable: true
      example:
        error:
          code: VALIDATION_ERROR
          message: Invalid request parameters
          details:
            field: nickname
            issue: Length must be between 2 and 20 characters
  responses:
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: BAD_REQUEST
              message: Invalid request parameters
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: An unexpected error occurred